Restructured video capture: Fake capture is now a part of uvc_capture. The user can select and switch from a normal capture to fake capture during runtime.
Pyuvc v.03 now has a new function to check if a device is already in use (is_accessible) with this we select not the first match but the first match that is not in use.
Multiple sessions of Pupil Capture can now be opened without trying to open cameras that are already in use from a running Pupil Capture session.
Pupil Capture process startup is now forced to be sequential.  This simplifies the binocular camera assignment logic.

Auto create capture has been removed from all player related plugins and player itself. Instead we use file capture directly.

The windows backend still needs to be changed to accommodate this restructuring.