set default y position of windows to 90 for Windows OS
If a user sets the system task bar to be located at the top of their screen, applicaiton window controls will be located underneath the task bar with old default y value settings for window_position_default and therefore not possible to move or close the application window. The current y value is set to 90 as the default Windows task bar height is approx 50 pixels in height and the window controls are approx 40 pixels in height.

Bigger question - why would Windows OS allow applicaitons to create windows under the task bar **or** why would Winodws OS let users drag and release a window control under the task bar in the first place?