Add workaround for timeline display issues
Timeline will convert lines to float32, but with PI we have unix epoch timestamps.
While they are in range of float32, the precision suffers greatly and will be in the order of days!
Thus by subtracting t0, we can shift the timestamps to a more reasonable range and thus increase float32 precision.

NOTE: Currently this will have no effect because of a bug with PI recordings from the Android side!