Add pts to Frames from file backend and cleanup FakeFrame a bit
When exporting with AV_Writer it is much easier to have variable pts if the Frames coming from the file-backend already have that information.
Luckily we know the pts of the frames when getting them, so I just attached that.
Needed to adjust FakeFrames as well and did a little bit of cleanup.