Refactor AV_Writer/JPEG_Writer
I kept the old classes and put an underscore _ in front of them in order to keep the git diff cleaner.
Otherwise git would try to match all kind of hunks and produce complete chaos.

Summary of changes:

1. Instead of two separate but overlapping classes (AV_Writer and JPEG_Writer) we now have an abstract base class AV_Writer and two specializations, MPEG_Writer and JPEG_Writer.
This makes more sense, as the JPEG_Writer also uses pyAV and the previous AV_Writer was only writing MPEG.

2. Extracting common code was challenging, as the data flow was very different between AV_Writer and JPEG_Writer, although both classes did inherently similar things.
The main point of generalization was to have an abstract method encode_frame(), which receives a frame-object and returns a number of Packets.
This is the main difference between both writers and lining that out made everything a lot clearer.

3. Audio writing functionality was apparently hacked into AV_Writer at some point, but was quite different from the rest of the flow.
I created a child class of MPEFG_Writer, called MPEG_Audio_Writer, which wraps additional audio encoding to separate audio from video.
Currently this is only used for world export.

4. All dependecy on the FPS has been removed from both writers, as this was only used to write fixed-rate PTS.

5. The writers now have two modes for PTS, controlled by the __init__ parameter frame_pts_timebase:
* If no timebase is set, frame-pts will be calculated from their timestamps.
* If a timebase is set, frames are expected to carry their own pts. The output stream will be created with the same timebase.