Always calculate PTS from timestamp
This is basically reverting some previous changes. I figured it makes no sense to try to keep existing pts on frames when exporting, mostly because PTS make absolutely no sense for multipart-recordings. Additionally this was already complex enough because time bases for the streams might be different when exporting. Removing PTS makes this simpler.

Changes made:
- Remove pts from all frames on file_backend as we don't need to transmit those anymore.
- Remove all frame_pts_timebase related logic in av_writer, as we now just always calculate PTS.
- Use start_time synced in av_writer to calculate PTS instead of first-frame timestamp
  - This required packing start_time_synced into the notification dict for recording
  - For export the time is read from info.csv instead