Fix issue with corners remaining stuck to the mouse pointer
Only consider how close the current pointer position is to the corner if it's not already being moved. If the corner is in the process of being moved, release it on mouse release no mater the distance to the pointer. This avoids situations where moving the corner very fast gets the UI into a state where the corner is stuck to the mouse pointer, but with an offset larger than the distance threshold, making it impossible to fixate.