Disable marker neighbor lookups for surface definition
This is very unstable of neighboring frames in the surrounding of the surface definition have flickering in the markers or in general have sets of markers that do not overlap much between frames.

This is a quick fix. The best would be to remove all the code that is responsible for doing the neighbor smoothing.