Correctly shutdown Service if port is already in use
The problem was that the shutdown of Service would be so fast that the eye processes were not started up completely.
Thus the cleanup at the end of Service was not received by the eye-processes yet and they would be kept alive.

Preventing the eye processes from starting in case of an initialization failure was way easier to implement and provides better UX than waiting for the eye processes to finish startup before shutting down Service.