Make PI raw gaze glob pattern more specific
This is important because there was an additional debug file introduced to PI recordings, called `gaze_right ps*_timestamps.npy` which (currently) only contains zeros.
With the previous pattern we intended to match all gaze files for multipart recordings, but now we would include these erroneous zeros, leading to jumps in the gaze data.