Always return Video to avoid container_idx confusion in fill_gaps=False situations
Issue: The lookup table is build with all videos (incl. invalid ones). If fill_gaps=False we mask the table to not include invalid videos. If we also filter the list of videos in this case, they won't match the container_idx in the masked lookup table anymore. Solution is to not filter the videos.