Implement Model2D outlier removal and refitting
This works the same way as the old `calibrate_2d.py` code did. However,
I fixed a 'bug' in the old version, where the reported errors where
essentially only half of the true errors. Since this was purely used for
filtering and reporting, this did not have any severe consequences. The
only change needed was to push the THRESHOLD from 35 to 70 in order to
keep the same behavior.

To avoid code duplication, I implemented this as recursive function.
This way we could also potentially run multiple iterations, although
this would probably overfit.