Allow replacing `by_class` unique plugins
Changes previous behavior where running `by_class` unique plugins could not be replaced.

This behaviour is especially problematic if `by_class` plugins are started by default. In this case, there would not be a possibility to change their configuration.

With the new behaviour, a new configuration can be loaded by restarting the plugin with the new config. The old plugin will be removed.