Fix MutableBisector.insert
MutableBisector calls numpy.insert on self.data.  The documentation states that it will attempt to convert the insertion value to the dtype of the array. https://numpy.org/doc/1.18/reference/generated/numpy.insert.html

This currently fails for post-hoc annotations and head pose tracking as self.data is not an object-array.