Remove unnecessary on_resize call
glfwSetWindowSize triggers a frame buffer change which will call on_resize with appropriate parameters.

This fixes an issue on Mac retina displays which caused the world window to flicker after clicking the "Reset window size" button as on_resize() was not called with the window frame buffer size but with the window size. On retina displays, these are different which cause the issue to appear.