Make camera model constructors consistent
This fixes a bug where loading dummy intrinsics from file would crash
because the Dummy_Camera constructor does not reasonably overwrite the
parent constructor. I moved name and resolution to from for all
constructors as these are the least likely to change and adjusted the
Dummy_Camera contructor to be consistent with the others.