Use packet instead of frame pts to check whether a frame was skipped
Pupil Player assumes that scene video packet and frame pts are equal and will warn if a new frame's pts does not conform with the target packet pts.

This commit yields packet and frame pts such that Player can compare the target packet pts to the actual packet pts from which the frame was decoded from.

For this to work well, there is the assumption that every packet yields one frame only.